requirements.txt
flask==2.0.2
rpi_ws281x==4.3.1
adafruit-circuitpython-neopixel==6.2.3
python-config==0.1.2
gpiod==1.5.0
